+++ apps/playground-vue/vite.config.ts
@@ -1,9 +1,10 @@
 import path from 'node:path'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import vueJsx from '@vitejs/plugin-vue-jsx'
 
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [vue(), vueJsx()],
   resolve: {
     alias: [
       {
@@ -20,8 +21,17 @@ export default defineConfig({
       }
     ]
   },
+  css: {
+    preprocessorOptions: {
+      scss: {
+        api: 'modern-compiler',
+        silenceDeprecations: ['legacy-js-api']
+      }
+    }
+  },
   server: {
     port: 8010,
+    strictPort: false,
     proxy: {
       '/api': {
         target: 'http://127.0.0.1:8008',